Populate OffsetIndex and ColumnIndex for each row_group
The statistics for each page in a ColumnChunk of a RowGroup are
added to the ColumnIndex structure. When a page is flushed to the
file, its location and offset of the first row is added to the
PageLocation structure of the Offset index. If a file is found to
have only one row_group when it is Finalized, the OffsetIndex and
ColumnIndex for each Column are written to the file (just before
footer) and the length and offsets of these indexes are populated
in the ColumnChunk. This ensures that the rage scans and point look
ups can skip pages based on these statistics while at the same time
scans without selective predicates do not incur any overhead. Space
efficiency is ensured by not populating parquet::Statistics in the
ColumnMeta when the statistics are written to the ColumnIndex.
Additionally, for SortingColumns, the ColumnIndex only contains the
min_values.

Testing: The populated index structures were deserialized from the
parquet file and the validity of the offsets and statistics were
verified.

Change-Id: Ied68f99c63971de76fd1d800e01c373bcff0f92a