Skip pages filtered by the index
The InitColumns function initializes a list of streams for each column.
Each stream either corresponds to the dictionary pages or the set of
filtered pages. These streams are combined in a single
ConcatenatedStreams object and passed to the parquet column reader. The
ScalarBaseColumnReader has a pointer to the FilteredPageInfos structure
to ensure that all the pages marked for reading by the IndexFilter are
read and reading the required pages marks the end of the RowGroup. In
the HdfsParquetScanner, the ValidateEndOfRowGroup function ensures that
number of rows read check is carried out only when the index filtering
of pages was not used.

Testing: (Manual) Two single column tables were created containing the
orderkey column of the tpch.lineitem table, one with a parquet index
and the other without. The outputs of various queries were compared to
ensure that index filtering produced consistent output.
The existing test_scanners.py test was used to validate the
changes don't break any of the existing flow  of code.

Change-Id: Ic7f7b3c553e9d25fccd4321c61ef20d51fdf516d